requirements.txt
asgiref==3.2.10
Django==3.1.2
django-environ==0.4.5
djangorestframework==3.12.1
environ==1.0
pathlib==1.0.1
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.4.1
